+++ koda_validate/_cruft.py
@@ -3,19 +3,19 @@ from __future__ import annotations
 from functools import partial
 from typing import Callable, Optional, Tuple, Union, overload
 
-from koda._generics import A, B, FailT, I, J
+from koda._generics import A, B, FailT, I, J, K
 from koda.result import Err, Result, Ok
 
 # simple alias allows to greatly reduce amount of characters required to define
 # validators
-from koda_validate._generics import Ret
+from koda_validate._generics import Ret, C, D, E, F, G, H
 
 _Validator = Callable[[A], Result[B, FailT]]
 
 
 def _flat_map_same_type_if_not_none(
-    fn: Optional[Callable[[A], Result[A, FailT]]],
-    r: Result[A, FailT],
+        fn: Optional[Callable[[A], Result[A, FailT]]],
+        r: Result[A, FailT],
 ) -> Result[A, FailT]:
     if fn is None:
         return r
@@ -25,163 +25,163 @@ def _flat_map_same_type_if_not_none(
 
 @overload
 def _validate_and_map(
-    r1: Result[A, FailT],
-    r2: Callable[[A], Ret],
-    *,
-    validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
+        r1: Result[A, FailT],
+        r2: Callable[[A], Ret],
+        *,
+        validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
 ) -> Result[Ret, Tuple[FailT, ...]]:
     ...
 
 
 @overload
 def _validate_and_map(
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Callable[[A, B], Ret],
-    *,
-    validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Callable[[A, B], Ret],
+        *,
+        validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
 ) -> Result[Ret, Tuple[FailT, ...]]:
     ...
 
 
 @overload
 def _validate_and_map(
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Callable[[A, B, C], Ret],
-    *,
-    validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Callable[[A, B, C], Ret],
+        *,
+        validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
 ) -> Result[Ret, Tuple[FailT, ...]]:
     ...
 
 
 @overload
 def _validate_and_map(
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Result[D, FailT],
-    r5: Callable[[A, B, C, D], Ret],
-    *,
-    validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Result[D, FailT],
+        r5: Callable[[A, B, C, D], Ret],
+        *,
+        validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
 ) -> Result[Ret, Tuple[FailT, ...]]:
     ...
 
 
 @overload
 def _validate_and_map(
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Result[D, FailT],
-    r5: Result[E, FailT],
-    r6: Callable[[A, B, C, D, E], Ret],
-    *,
-    validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Result[D, FailT],
+        r5: Result[E, FailT],
+        r6: Callable[[A, B, C, D, E], Ret],
+        *,
+        validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
 ) -> Result[Ret, Tuple[FailT, ...]]:
     ...
 
 
 @overload
 def _validate_and_map(
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Result[D, FailT],
-    r5: Result[E, FailT],
-    r6: Result[F, FailT],
-    r7: Callable[[A, B, C, D, E, F], Ret],
-    *,
-    validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Result[D, FailT],
+        r5: Result[E, FailT],
+        r6: Result[F, FailT],
+        r7: Callable[[A, B, C, D, E, F], Ret],
+        *,
+        validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
 ) -> Result[Ret, Tuple[FailT, ...]]:
     ...
 
 
 @overload
 def _validate_and_map(
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Result[D, FailT],
-    r5: Result[E, FailT],
-    r6: Result[F, FailT],
-    r7: Result[G, FailT],
-    r8: Callable[[A, B, C, D, E, F, G], Ret],
-    *,
-    validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Result[D, FailT],
+        r5: Result[E, FailT],
+        r6: Result[F, FailT],
+        r7: Result[G, FailT],
+        r8: Callable[[A, B, C, D, E, F, G], Ret],
+        *,
+        validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
 ) -> Result[Ret, Tuple[FailT, ...]]:
     ...
 
 
 @overload
 def _validate_and_map(
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Result[D, FailT],
-    r5: Result[E, FailT],
-    r6: Result[F, FailT],
-    r7: Result[G, FailT],
-    r8: Result[H, FailT],
-    r9: Callable[[A, B, C, D, E, F, G, H], Ret],
-    *,
-    validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Result[D, FailT],
+        r5: Result[E, FailT],
+        r6: Result[F, FailT],
+        r7: Result[G, FailT],
+        r8: Result[H, FailT],
+        r9: Callable[[A, B, C, D, E, F, G, H], Ret],
+        *,
+        validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
 ) -> Result[Ret, Tuple[FailT, ...]]:
     ...
 
 
 @overload
 def _validate_and_map(
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Result[D, FailT],
-    r5: Result[E, FailT],
-    r6: Result[F, FailT],
-    r7: Result[G, FailT],
-    r8: Result[H, FailT],
-    r9: Result[I, FailT],
-    r10: Callable[[A, B, C, D, E, F, G, H, I], Ret],
-    *,
-    validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Result[D, FailT],
+        r5: Result[E, FailT],
+        r6: Result[F, FailT],
+        r7: Result[G, FailT],
+        r8: Result[H, FailT],
+        r9: Result[I, FailT],
+        r10: Callable[[A, B, C, D, E, F, G, H, I], Ret],
+        *,
+        validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
 ) -> Result[Ret, Tuple[FailT, ...]]:
     ...
 
 
 @overload
 def _validate_and_map(
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Result[D, FailT],
-    r5: Result[E, FailT],
-    r6: Result[F, FailT],
-    r7: Result[G, FailT],
-    r8: Result[H, FailT],
-    r9: Result[I, FailT],
-    r10: Result[J, FailT],
-    r11: Callable[[A, B, C, D, E, F, G, H, I, J], Ret],
-    *,
-    validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Result[D, FailT],
+        r5: Result[E, FailT],
+        r6: Result[F, FailT],
+        r7: Result[G, FailT],
+        r8: Result[H, FailT],
+        r9: Result[I, FailT],
+        r10: Result[J, FailT],
+        r11: Callable[[A, B, C, D, E, F, G, H, I, J], Ret],
+        *,
+        validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
 ) -> Result[Ret, Tuple[FailT, ...]]:
     ...
 
 
 def _validate_and_map(
-    r1: Result[A, FailT],
-    r2: Callable[[A], Ret] | Result[B, FailT],
-    r3: None | Callable[[A, B], Ret] | Result[C, FailT] = None,
-    r4: None | Callable[[A, B, C], Ret] | Result[D, FailT] = None,
-    r5: None | Callable[[A, B, C, D], Ret] | Result[E, FailT] = None,
-    r6: None | Callable[[A, B, C, D, E], Ret] | Result[F, FailT] = None,
-    r7: None | Callable[[A, B, C, D, E, F], Ret] | Result[G, FailT] = None,
-    r8: None | Callable[[A, B, C, D, E, F, G], Ret] | Result[H, FailT] = None,
-    r9: None | Callable[[A, B, C, D, E, F, G, H], Ret] | Result[I, FailT] = None,
-    r10: None | Callable[[A, B, C, D, E, F, G, H, I], Ret] | Result[J, FailT] = None,
-    r11: None | Callable[[A, B, C, D, E, F, G, H, I, J], Ret] = None,
-    *,
-    validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
+        r1: Result[A, FailT],
+        r2: Callable[[A], Ret] | Result[B, FailT],
+        r3: None | Callable[[A, B], Ret] | Result[C, FailT] = None,
+        r4: None | Callable[[A, B, C], Ret] | Result[D, FailT] = None,
+        r5: None | Callable[[A, B, C, D], Ret] | Result[E, FailT] = None,
+        r6: None | Callable[[A, B, C, D, E], Ret] | Result[F, FailT] = None,
+        r7: None | Callable[[A, B, C, D, E, F], Ret] | Result[G, FailT] = None,
+        r8: None | Callable[[A, B, C, D, E, F, G], Ret] | Result[H, FailT] = None,
+        r9: None | Callable[[A, B, C, D, E, F, G, H], Ret] | Result[I, FailT] = None,
+        r10: None | Callable[[A, B, C, D, E, F, G, H, I], Ret] | Result[J, FailT] = None,
+        r11: None | Callable[[A, B, C, D, E, F, G, H, I, J], Ret] = None,
+        *,
+        validate_object: Optional[Callable[[Ret], Result[Ret, Tuple[FailT, ...]]]] = None
 ) -> Result[Ret, Tuple[FailT, ...]]:
     """
     See overloaded signatures above for better idea of what's happening here
@@ -277,60 +277,60 @@ def _validate_and_map(
 
 @overload
 def _chain(
-    fn1: _Validator[A, B, FailT], fn2: _Validator[B, C, FailT]
+        fn1: _Validator[A, B, FailT], fn2: _Validator[B, C, FailT]
 ) -> _Validator[A, C, FailT]:
     ...
 
 
 @overload
 def _chain(
-    fn1: _Validator[A, B, FailT],
-    fn2: _Validator[B, C, FailT],
-    fn3: _Validator[C, D, FailT],
+        fn1: _Validator[A, B, FailT],
+        fn2: _Validator[B, C, FailT],
+        fn3: _Validator[C, D, FailT],
 ) -> _Validator[A, D, FailT]:
     ...
 
 
 @overload
 def _chain(
-    fn1: _Validator[A, B, FailT],
-    fn2: _Validator[B, C, FailT],
-    fn3: _Validator[C, D, FailT],
-    fn4: _Validator[D, E, FailT],
+        fn1: _Validator[A, B, FailT],
+        fn2: _Validator[B, C, FailT],
+        fn3: _Validator[C, D, FailT],
+        fn4: _Validator[D, E, FailT],
 ) -> _Validator[A, E, FailT]:
     ...
 
 
 @overload
 def _chain(
-    fn1: _Validator[A, B, FailT],
-    fn2: _Validator[B, C, FailT],
-    fn3: _Validator[C, D, FailT],
-    fn4: _Validator[D, E, FailT],
-    fn5: _Validator[E, F, FailT],
+        fn1: _Validator[A, B, FailT],
+        fn2: _Validator[B, C, FailT],
+        fn3: _Validator[C, D, FailT],
+        fn4: _Validator[D, E, FailT],
+        fn5: _Validator[E, F, FailT],
 ) -> _Validator[A, F, FailT]:
     ...
 
 
 @overload
 def _chain(
-    fn1: _Validator[A, B, FailT],
-    fn2: _Validator[B, C, FailT],
-    fn3: _Validator[C, D, FailT],
-    fn4: _Validator[D, E, FailT],
-    fn5: _Validator[E, F, FailT],
-    fn6: _Validator[F, G, FailT],
+        fn1: _Validator[A, B, FailT],
+        fn2: _Validator[B, C, FailT],
+        fn3: _Validator[C, D, FailT],
+        fn4: _Validator[D, E, FailT],
+        fn5: _Validator[E, F, FailT],
+        fn6: _Validator[F, G, FailT],
 ) -> _Validator[A, G, FailT]:
     ...
 
 
 def _chain(
-    fn1: _Validator[A, B, FailT],
-    fn2: _Validator[B, C, FailT],
-    fn3: Optional[_Validator[C, D, FailT]] = None,
-    fn4: Optional[_Validator[D, E, FailT]] = None,
-    fn5: Optional[_Validator[E, F, FailT]] = None,
-    fn6: Optional[_Validator[F, G, FailT]] = None,
+        fn1: _Validator[A, B, FailT],
+        fn2: _Validator[B, C, FailT],
+        fn3: Optional[_Validator[C, D, FailT]] = None,
+        fn4: Optional[_Validator[D, E, FailT]] = None,
+        fn5: Optional[_Validator[E, F, FailT]] = None,
+        fn6: Optional[_Validator[F, G, FailT]] = None,
 ) -> Callable[
     [A],
     Union[
@@ -342,7 +342,7 @@ def _chain(
     ],
 ]:
     def inner(
-        val: A,
+            val: A,
     ) -> Union[
         Result[C, FailT],
         Result[D, FailT],
@@ -361,18 +361,18 @@ def _chain(
         else:
             return (
                 fn1(val)
-                .flat_map(fn2)
-                .flat_map(fn3)
-                .flat_map(fn4)
-                .flat_map(fn5)
-                .flat_map(fn6)
+                    .flat_map(fn2)
+                    .flat_map(fn3)
+                    .flat_map(fn4)
+                    .flat_map(fn5)
+                    .flat_map(fn6)
             )
 
     return inner
 
 
 def _validate1_helper(
-    state: Result[Callable[[A], B], Tuple[FailT, ...]], r: Result[A, FailT]
+        state: Result[Callable[[A], B], Tuple[FailT, ...]], r: Result[A, FailT]
 ) -> Result[B, Tuple[FailT, ...]]:
     if isinstance(r, Err):
         if isinstance(state, Err):
@@ -387,9 +387,9 @@ def _validate1_helper(
 
 
 def _validate2_helper(
-    state: Result[Callable[[A, B], C], Tuple[FailT, ...]],
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
+        state: Result[Callable[[A, B], C], Tuple[FailT, ...]],
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
 ) -> Result[C, Tuple[FailT, ...]]:
     if isinstance(r1, Err):
         if isinstance(state, Err):
@@ -408,10 +408,10 @@ def _validate2_helper(
 
 
 def _validate3_helper(
-    state: Result[Callable[[A, B, C], D], Tuple[FailT, ...]],
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
+        state: Result[Callable[[A, B, C], D], Tuple[FailT, ...]],
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
 ) -> Result[D, Tuple[FailT, ...]]:
     if isinstance(r1, Err):
         if isinstance(state, Err):
@@ -430,11 +430,11 @@ def _validate3_helper(
 
 
 def _validate4_helper(
-    state: Result[Callable[[A, B, C, D], E], Tuple[FailT, ...]],
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Result[D, FailT],
+        state: Result[Callable[[A, B, C, D], E], Tuple[FailT, ...]],
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Result[D, FailT],
 ) -> Result[E, Tuple[FailT, ...]]:
     if isinstance(r1, Err):
         if isinstance(state, Err):
@@ -453,12 +453,12 @@ def _validate4_helper(
 
 
 def _validate5_helper(
-    state: Result[Callable[[A, B, C, D, E], F], Tuple[FailT, ...]],
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Result[D, FailT],
-    r5: Result[E, FailT],
+        state: Result[Callable[[A, B, C, D, E], F], Tuple[FailT, ...]],
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Result[D, FailT],
+        r5: Result[E, FailT],
 ) -> Result[F, Tuple[FailT, ...]]:
     if isinstance(r1, Err):
         if isinstance(state, Err):
@@ -477,13 +477,13 @@ def _validate5_helper(
 
 
 def _validate6_helper(
-    state: Result[Callable[[A, B, C, D, E, F], G], Tuple[FailT, ...]],
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Result[D, FailT],
-    r5: Result[E, FailT],
-    r6: Result[F, FailT],
+        state: Result[Callable[[A, B, C, D, E, F], G], Tuple[FailT, ...]],
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Result[D, FailT],
+        r5: Result[E, FailT],
+        r6: Result[F, FailT],
 ) -> Result[G, Tuple[FailT, ...]]:
     if isinstance(r1, Err):
         if isinstance(state, Err):
@@ -502,14 +502,14 @@ def _validate6_helper(
 
 
 def _validate7_helper(
-    state: Result[Callable[[A, B, C, D, E, F, G], H], Tuple[FailT, ...]],
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Result[D, FailT],
-    r5: Result[E, FailT],
-    r6: Result[F, FailT],
-    r7: Result[G, FailT],
+        state: Result[Callable[[A, B, C, D, E, F, G], H], Tuple[FailT, ...]],
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Result[D, FailT],
+        r5: Result[E, FailT],
+        r6: Result[F, FailT],
+        r7: Result[G, FailT],
 ) -> Result[H, Tuple[FailT, ...]]:
     if isinstance(r1, Err):
         if isinstance(state, Err):
@@ -528,15 +528,15 @@ def _validate7_helper(
 
 
 def _validate8_helper(
-    state: Result[Callable[[A, B, C, D, E, F, G, H], I], Tuple[FailT, ...]],
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Result[D, FailT],
-    r5: Result[E, FailT],
-    r6: Result[F, FailT],
-    r7: Result[G, FailT],
-    r8: Result[H, FailT],
+        state: Result[Callable[[A, B, C, D, E, F, G, H], I], Tuple[FailT, ...]],
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Result[D, FailT],
+        r5: Result[E, FailT],
+        r6: Result[F, FailT],
+        r7: Result[G, FailT],
+        r8: Result[H, FailT],
 ) -> Result[I, Tuple[FailT, ...]]:
     if isinstance(r1, Err):
         if isinstance(state, Err):
@@ -555,16 +555,16 @@ def _validate8_helper(
 
 
 def _validate9_helper(
-    state: Result[Callable[[A, B, C, D, E, F, G, H, I], J], Tuple[FailT, ...]],
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Result[D, FailT],
-    r5: Result[E, FailT],
-    r6: Result[F, FailT],
-    r7: Result[G, FailT],
-    r8: Result[H, FailT],
-    r9: Result[I, FailT],
+        state: Result[Callable[[A, B, C, D, E, F, G, H, I], J], Tuple[FailT, ...]],
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Result[D, FailT],
+        r5: Result[E, FailT],
+        r6: Result[F, FailT],
+        r7: Result[G, FailT],
+        r8: Result[H, FailT],
+        r9: Result[I, FailT],
 ) -> Result[J, Tuple[FailT, ...]]:
     if isinstance(r1, Err):
         if isinstance(state, Err):
@@ -583,17 +583,17 @@ def _validate9_helper(
 
 
 def _validate10_helper(
-    state: Result[Callable[[A, B, C, D, E, F, G, H, I, J], K], Tuple[FailT, ...]],
-    r1: Result[A, FailT],
-    r2: Result[B, FailT],
-    r3: Result[C, FailT],
-    r4: Result[D, FailT],
-    r5: Result[E, FailT],
-    r6: Result[F, FailT],
-    r7: Result[G, FailT],
-    r8: Result[H, FailT],
-    r9: Result[I, FailT],
-    r10: Result[J, FailT],
+        state: Result[Callable[[A, B, C, D, E, F, G, H, I, J], K], Tuple[FailT, ...]],
+        r1: Result[A, FailT],
+        r2: Result[B, FailT],
+        r3: Result[C, FailT],
+        r4: Result[D, FailT],
+        r5: Result[E, FailT],
+        r6: Result[F, FailT],
+        r7: Result[G, FailT],
+        r8: Result[H, FailT],
+        r9: Result[I, FailT],
+        r10: Result[J, FailT],
 ) -> Result[K, Tuple[FailT, ...]]:
     if isinstance(r1, Err):
         if isinstance(state, Err):
@@ -612,15 +612,6 @@ def _validate10_helper(
 
 
 from typing import Tuple, TypeVar, Union, overload
-
-A = TypeVar("A")
-B = TypeVar("B")
-C = TypeVar("C")
-D = TypeVar("D")
-E = TypeVar("E")
-F = TypeVar("F")
-G = TypeVar("G")
-H = TypeVar("H")
 
 
 class _NotSet:
@@ -664,27 +655,27 @@ def _typed_tuple(v1: A, v2: B, v3: C, v4: D, v5: E, v6: F) -> Tuple[A, B, C, D, 
 
 @overload
 def _typed_tuple(
-    v1: A, v2: B, v3: C, v4: D, v5: E, v6: F, v7: G
+        v1: A, v2: B, v3: C, v4: D, v5: E, v6: F, v7: G
 ) -> Tuple[A, B, C, D, E, F, G]:
     ...
 
 
 @overload
 def _typed_tuple(
-    v1: A, v2: B, v3: C, v4: D, v5: E, v6: F, v7: G, v8: H
+        v1: A, v2: B, v3: C, v4: D, v5: E, v6: F, v7: G, v8: H
 ) -> Tuple[A, B, C, D, E, F, G, H]:
     ...
 
 
 def _typed_tuple(
-    v1: A,
-    v2: _Settable[B] = _not_set,
-    v3: _Settable[C] = _not_set,
-    v4: _Settable[D] = _not_set,
-    v5: _Settable[E] = _not_set,
-    v6: _Settable[F] = _not_set,
-    v7: _Settable[G] = _not_set,
-    v8: _Settable[H] = _not_set,
+        v1: A,
+        v2: _Settable[B] = _not_set,
+        v3: _Settable[C] = _not_set,
+        v4: _Settable[D] = _not_set,
+        v5: _Settable[E] = _not_set,
+        v6: _Settable[F] = _not_set,
+        v7: _Settable[G] = _not_set,
+        v8: _Settable[H] = _not_set,
 ) -> Union[
     Tuple[A],
     Tuple[A, B],
