+++ tests/test_validation.py
@@ -28,18 +28,18 @@ from koda_validate.validation import (
     MinProperties,
     MultipleOf,
     NotBlank,
-    Null,
+    null,
     Nullable,
-    Obj1,
-    Obj2,
-    Obj3,
-    Obj4,
-    Obj5,
-    Obj6,
-    Obj7,
-    Obj8,
-    Obj9,
-    Obj10,
+    Obj1Prop,
+    Obj2Props,
+    Obj3Props,
+    Obj4Props,
+    Obj5Props,
+    Obj6Props,
+    Obj7Props,
+    Obj8Props,
+    Obj9Props,
+    Obj10Props,
     OneOf2,
     OneOf3,
     RegexValidator,
@@ -52,7 +52,7 @@ from koda_validate.validation import (
     maybe_prop,
     prop,
     unique_items,
-    unwrap_jsonable, Decimal,
+    unwrap_jsonable, Decimal, validate_and_map,
 )
 from koda.maybe import Just, Maybe, nothing
 from koda.result import Err, Result, Ok
@@ -133,11 +133,11 @@ def test_date() -> None:
 
 
 def test_null() -> None:
-    assert Null("a string") == err_list("expected null")
+    assert null("a string") == err_list("expected null")
 
-    assert Null(None) == Ok(None)
+    assert null(None) == Ok(None)
 
-    assert Null(False) == err_list("expected null")
+    assert null(False) == err_list("expected null")
 
 
 def test_integer() -> None:
@@ -435,7 +435,10 @@ def test_obj_1() -> None:
     class Person:
         name: str
 
-    validator = Obj1(prop("name", String()), into=Person)
+    validator = Obj1Prop(
+        prop("name", String()),
+        into=Person
+    )
 
     assert validator("not a dict") == err({"__object__": ["expected an object"]})
 
@@ -456,7 +459,7 @@ def test_obj_2() -> None:
         name: str
         age: Maybe[int]
 
-    validator = Obj2(prop("name", String()), maybe_prop("age", Integer()), into=Person)
+    validator = Obj2Props(prop("name", String()), maybe_prop("age", Integer()), into=Person)
 
     assert validator("not a dict") == err({"__object__": ["expected an object"]})
 
@@ -481,7 +484,7 @@ def test_obj_3() -> None:
         last_name: str
         age: int
 
-    validator = Obj3(
+    validator = Obj3Props(
         prop("first_name", String()),
         prop("last_name", String()),
         prop("age", Integer()),
@@ -522,7 +525,7 @@ def test_obj_4() -> None:
         age: int
         eye_color: str
 
-    validator = Obj4(
+    validator = Obj4Props(
         prop("first_name", String()),
         prop("last_name", String()),
         prop("age", Integer()),
@@ -551,7 +554,7 @@ def test_obj_5() -> None:
         eye_color: str
         can_fly: bool
 
-    validator = Obj5(
+    validator = Obj5Props(
         prop("first_name", String()),
         prop("last_name", String()),
         prop("age", Integer()),
@@ -600,7 +603,7 @@ def test_obj_6() -> None:
         can_fly: bool
         fingers: float
 
-    validator = Obj6(
+    validator = Obj6Props(
         prop("first_name", String()),
         prop("last_name", String()),
         prop("age", Integer()),
@@ -638,7 +641,7 @@ def test_obj_7() -> None:
         fingers: float
         toes: float
 
-    validator = Obj7(
+    validator = Obj7Props(
         prop("first_name", String()),
         prop("last_name", String()),
         prop("age", Integer()),
@@ -679,7 +682,7 @@ def test_obj_8() -> None:
         toes: float
         favorite_color: Maybe[str]
 
-    validator = Obj8(
+    validator = Obj8Props(
         prop("first_name", String()),
         prop("last_name", String()),
         prop("age", Integer()),
@@ -755,7 +758,7 @@ def test_obj_9() -> None:
         favorite_color: Maybe[str]
         requires_none: None
 
-    validator = Obj9(
+    validator = Obj9Props(
         prop("first_name", String()),
         prop("last_name", String()),
         prop("age", Integer()),
@@ -764,7 +767,7 @@ def test_obj_9() -> None:
         prop("number_of_fingers", Float()),
         prop("number of toes", Float()),
         maybe_prop("favorite_color", String()),
-        prop("requires_none", Null),
+        prop("requires_none", null),
         into=Person,
         validate_object=_nobody_named_jones_has_brown_eyes,
     )
@@ -803,7 +806,7 @@ def test_obj_10() -> None:
         requires_none: None
         something_else: List[str]
 
-    validator = Obj10(
+    validator = Obj10Props(
         prop("first_name", String()),
         prop("last_name", String()),
         prop("age", Integer()),
@@ -812,7 +815,7 @@ def test_obj_10() -> None:
         prop("number_of_fingers", Float()),
         prop("number of toes", Float()),
         maybe_prop("favorite_color", String()),
-        prop("requires_none", Null),
+        prop("requires_none", null),
         prop("favorite_books", ArrayOf(String())),
         into=Person,
         validate_object=_nobody_named_jones_has_brown_eyes,
@@ -950,7 +953,7 @@ def deserialize_and_validate_tests() -> None:
         name: str
         age: int
 
-    validator = Obj2(prop("name", String()), prop("int", Integer()), into=Person)
+    validator = Obj2Props(prop("name", String()), prop("int", Integer()), into=Person)
 
     assert deserialize_and_validate(validator, "") == Err(
         {"invalid type": ["expected an object"]}
@@ -969,10 +972,10 @@ def test_lazy() -> None:
         val: int
         next: Maybe["TestNonEmptyList"]  # noqa: F821
 
-    def recur_tnel() -> Obj2[int, Maybe[TestNonEmptyList], TestNonEmptyList]:
+    def recur_tnel() -> Obj2Props[int, Maybe[TestNonEmptyList], TestNonEmptyList]:
         return nel_validator
 
-    nel_validator: Obj2[int, Maybe[TestNonEmptyList], TestNonEmptyList] = Obj2(
+    nel_validator: Obj2Props[int, Maybe[TestNonEmptyList], TestNonEmptyList] = Obj2Props(
         prop("val", Integer()),
         maybe_prop("next", Lazy(recur_tnel)),
         into=TestNonEmptyList,
@@ -1006,3 +1009,628 @@ def test_multiple_of() -> None:
     assert MultipleOf(5)(10) == Ok(10)
     assert MultipleOf(5)(11) == err("expected multiple of 5")
     assert MultipleOf(2.2)(4.40) == Ok(4.40)
+
+
+def test_map_1() -> None:
+    def string_thing(s: str) -> str:
+        return f"valid {s}"
+
+    assert (
+            validate_and_map(
+                Err("just returns failure"),
+                string_thing,
+            )
+            == Err(("just returns failure",))
+    )
+
+    assert validate_and_map(Ok("hooray"), string_thing) == Ok("valid hooray")
+
+
+def test_map_2() -> None:
+    @dataclass
+    class Person2:
+        name: str
+        age: int
+
+    assert validate_and_map(Err("invalid name"), Err("invalid age"), Person2) == Err(
+        ("invalid name", "invalid age")
+    )
+
+    assert validate_and_map(Ok("Bob"), Err("invalid age"), Person2) == Err(
+        ("invalid age",)
+    )
+
+    assert validate_and_map(Ok("Bob"), Ok(25), Person2) == Ok(
+        Person2(name="Bob", age=25)
+    )
+
+
+def test_map_3() -> None:
+    @dataclass
+    class Person:
+        first_name: str
+        last_name: str
+        age: int
+
+    assert validate_and_map(
+        Err("invalid first name"), Err("invalid last name"), Err("invalid age"), Person
+    ) == Err(("invalid first name", "invalid last name", "invalid age"))
+
+    assert validate_and_map(
+        Err("invalid first name"), Err("invalid last name"), Ok(25), Person
+    ) == Err(("invalid first name", "invalid last name"))
+
+    assert validate_and_map(
+        Err("invalid first name"), Ok("Smith"), Ok(25), Person
+    ) == Err(("invalid first name",))
+
+    assert validate_and_map(Ok("John"), Ok("Doe"), Ok(25), Person) == Ok(
+        Person(first_name="John", last_name="Doe", age=25)
+    )
+
+
+def test_map_4() -> None:
+    @dataclass
+    class Person:
+        first_name: str
+        last_name: str
+        age: int
+        eye_color: str
+
+    assert validate_and_map(
+        Err("invalid first name"),
+        Err("invalid last name"),
+        Err("invalid age"),
+        Err("invalid eye color"),
+        Person,
+    ) == Err(
+        ("invalid first name", "invalid last name", "invalid age", "invalid eye color")
+    )
+
+    assert (
+            validate_and_map(
+                Err("invalid first name"),
+                Err("invalid last name"),
+                Ok(25),
+                Err("invalid eye color"),
+                Person,
+            )
+            == Err(("invalid first name", "invalid last name", "invalid eye color"))
+    )
+
+    assert validate_and_map(
+        Err("invalid first name"), Ok("Smith"), Ok(25), Err("invalid eye color"), Person
+    ) == Err(("invalid first name", "invalid eye color"))
+
+    assert validate_and_map(Ok("John"), Ok("Doe"), Ok(25), Ok("brown"), Person) == Ok(
+        Person(first_name="John", last_name="Doe", age=25, eye_color="brown")
+    )
+
+
+def test_map_5() -> None:
+    @dataclass
+    class Person:
+        first_name: str
+        last_name: str
+        age: int
+        eye_color: str
+        can_fly: bool
+
+    assert validate_and_map(
+        Err("invalid first name"),
+        Err("invalid last name"),
+        Err("invalid age"),
+        Err("invalid eye color"),
+        Err("invalid -- people can't fly"),
+        Person,
+    ) == Err(
+        (
+            "invalid first name",
+            "invalid last name",
+            "invalid age",
+            "invalid eye color",
+            "invalid -- people can't fly",
+        )
+    )
+
+    assert (
+            validate_and_map(
+                Err("invalid first name"),
+                Err("invalid last name"),
+                Ok(25),
+                Err("invalid eye color"),
+                Ok(False),
+                Person,
+            )
+            == Err(("invalid first name", "invalid last name", "invalid eye color"))
+    )
+
+    assert (
+            validate_and_map(
+                Err("invalid first name"),
+                Ok("Smith"),
+                Ok(25),
+                Err("invalid eye color"),
+                Ok(False),
+                Person,
+            )
+            == Err(("invalid first name", "invalid eye color"))
+    )
+
+    assert validate_and_map(
+        Ok("John"), Ok("Doe"), Ok(25), Ok("brown"), Ok(False), Person
+    ) == Ok(
+        Person(
+            first_name="John", last_name="Doe", age=25, eye_color="brown", can_fly=False
+        )
+    )
+
+
+def test_map_6() -> None:
+    class SwimmingLevel(Enum):
+        BEGINNER = 1
+        INTERMEDIATE = 2
+        ADVANCED = 3
+
+    @dataclass
+    class Person:
+        first_name: str
+        last_name: str
+        age: int
+        eye_color: str
+        can_fly: bool
+        swimming_level: SwimmingLevel
+
+    assert validate_and_map(
+        Err("invalid first name"),
+        Err("invalid last name"),
+        Err("invalid age"),
+        Err("invalid eye color"),
+        Err("invalid -- people can't fly"),
+        Err("invalid -- beginners not allowed"),
+        Person,
+    ) == Err(
+        (
+            "invalid first name",
+            "invalid last name",
+            "invalid age",
+            "invalid eye color",
+            "invalid -- people can't fly",
+            "invalid -- beginners not allowed",
+        )
+    )
+
+    assert (
+            validate_and_map(
+                Err("invalid first name"),
+                Err("invalid last name"),
+                Ok(25),
+                Err("invalid eye color"),
+                Ok(False),
+                Ok(SwimmingLevel.ADVANCED),
+                Person,
+            )
+            == Err(("invalid first name", "invalid last name", "invalid eye color"))
+    )
+
+    assert (
+            validate_and_map(
+                Err("invalid first name"),
+                Ok("Smith"),
+                Ok(25),
+                Err("invalid eye color"),
+                Ok(False),
+                Ok(SwimmingLevel.ADVANCED),
+                Person,
+            )
+            == Err(("invalid first name", "invalid eye color"))
+    )
+
+    assert validate_and_map(
+        Ok("John"),
+        Ok("Doe"),
+        Ok(25),
+        Ok("brown"),
+        Ok(False),
+        Ok(SwimmingLevel.ADVANCED),
+        Person,
+    ) == Ok(
+        Person(
+            first_name="John",
+            last_name="Doe",
+            age=25,
+            eye_color="brown",
+            can_fly=False,
+            swimming_level=SwimmingLevel.ADVANCED,
+        )
+    )
+
+
+def test_map_7() -> None:
+    class SwimmingLevel(Enum):
+        BEGINNER = 1
+        INTERMEDIATE = 2
+        ADVANCED = 3
+
+    @dataclass
+    class Person:
+        first_name: str
+        last_name: str
+        age: int
+        eye_color: str
+        can_fly: bool
+        swimming_level: SwimmingLevel
+        number_of_fingers: float
+
+    assert validate_and_map(
+        Err("invalid first name"),
+        Err("invalid last name"),
+        Err("invalid age"),
+        Err("invalid eye color"),
+        Err("invalid -- people can't fly"),
+        Err("invalid -- beginners not allowed"),
+        Err("invalid number of fingers"),
+        Person,
+    ) == Err(
+        (
+            "invalid first name",
+            "invalid last name",
+            "invalid age",
+            "invalid eye color",
+            "invalid -- people can't fly",
+            "invalid -- beginners not allowed",
+            "invalid number of fingers",
+        )
+    )
+
+    assert (
+            validate_and_map(
+                Err("invalid first name"),
+                Err("invalid last name"),
+                Ok(25),
+                Err("invalid eye color"),
+                Ok(False),
+                Ok(SwimmingLevel.ADVANCED),
+                Ok(9.5),
+                Person,
+            )
+            == Err(("invalid first name", "invalid last name", "invalid eye color"))
+    )
+
+    assert (
+            validate_and_map(
+                Err("invalid first name"),
+                Ok("Smith"),
+                Ok(25),
+                Err("invalid eye color"),
+                Ok(False),
+                Ok(SwimmingLevel.ADVANCED),
+                Ok(9.5),
+                Person,
+            )
+            == Err(("invalid first name", "invalid eye color"))
+    )
+
+    assert validate_and_map(
+        Ok("John"),
+        Ok("Doe"),
+        Ok(25),
+        Ok("brown"),
+        Ok(False),
+        Ok(SwimmingLevel.ADVANCED),
+        Ok(9.5),
+        Person,
+    ) == Ok(
+        Person(
+            first_name="John",
+            last_name="Doe",
+            age=25,
+            eye_color="brown",
+            can_fly=False,
+            swimming_level=SwimmingLevel.ADVANCED,
+            number_of_fingers=9.5,
+        )
+    )
+
+
+def test_map_8() -> None:
+    class SwimmingLevel(Enum):
+        BEGINNER = 1
+        INTERMEDIATE = 2
+        ADVANCED = 3
+
+    @dataclass
+    class Person:
+        first_name: str
+        last_name: str
+        age: int
+        eye_color: str
+        can_fly: bool
+        swimming_level: SwimmingLevel
+        number_of_fingers: float
+        number_of_toes: float
+
+    assert validate_and_map(
+        Err("invalid first name"),
+        Err("invalid last name"),
+        Err("invalid age"),
+        Err("invalid eye color"),
+        Err("invalid -- people can't fly"),
+        Err("invalid -- beginners not allowed"),
+        Err("invalid number of fingers"),
+        Err("invalid number of toes"),
+        Person,
+    ) == Err(
+        (
+            "invalid first name",
+            "invalid last name",
+            "invalid age",
+            "invalid eye color",
+            "invalid -- people can't fly",
+            "invalid -- beginners not allowed",
+            "invalid number of fingers",
+            "invalid number of toes",
+        )
+    )
+
+    assert (
+            validate_and_map(
+                Err("invalid first name"),
+                Err("invalid last name"),
+                Ok(25),
+                Err("invalid eye color"),
+                Ok(False),
+                Ok(SwimmingLevel.ADVANCED),
+                Ok(9.5),
+                Ok(3.1),
+                Person,
+            )
+            == Err(("invalid first name", "invalid last name", "invalid eye color"))
+    )
+
+    assert (
+            validate_and_map(
+                Err("invalid first name"),
+                Ok("Smith"),
+                Ok(25),
+                Err("invalid eye color"),
+                Ok(False),
+                Ok(SwimmingLevel.ADVANCED),
+                Ok(9.5),
+                Ok(3.1),
+                Person,
+            )
+            == Err(("invalid first name", "invalid eye color"))
+    )
+
+    assert validate_and_map(
+        Ok("John"),
+        Ok("Doe"),
+        Ok(25),
+        Ok("brown"),
+        Ok(False),
+        Ok(SwimmingLevel.ADVANCED),
+        Ok(9.5),
+        Ok(3.1),
+        Person,
+    ) == Ok(
+        Person(
+            first_name="John",
+            last_name="Doe",
+            age=25,
+            eye_color="brown",
+            can_fly=False,
+            swimming_level=SwimmingLevel.ADVANCED,
+            number_of_fingers=9.5,
+            number_of_toes=3.1,
+        )
+    )
+
+
+def test_map_9() -> None:
+    class SwimmingLevel(Enum):
+        BEGINNER = 1
+        INTERMEDIATE = 2
+        ADVANCED = 3
+
+    @dataclass
+    class Person:
+        first_name: str
+        last_name: str
+        age: int
+        eye_color: str
+        can_fly: bool
+        swimming_level: SwimmingLevel
+        number_of_fingers: float
+        number_of_toes: float
+        country: str
+
+    assert validate_and_map(
+        Err("invalid first name"),
+        Err("invalid last name"),
+        Err("invalid age"),
+        Err("invalid eye color"),
+        Err("invalid -- people can't fly"),
+        Err("invalid -- beginners not allowed"),
+        Err("invalid number of fingers"),
+        Err("invalid number of toes"),
+        Err("invalid country"),
+        Person,
+    ) == Err(
+        (
+            "invalid first name",
+            "invalid last name",
+            "invalid age",
+            "invalid eye color",
+            "invalid -- people can't fly",
+            "invalid -- beginners not allowed",
+            "invalid number of fingers",
+            "invalid number of toes",
+            "invalid country",
+        )
+    )
+
+    assert (
+            validate_and_map(
+                Err("invalid first name"),
+                Err("invalid last name"),
+                Ok(25),
+                Err("invalid eye color"),
+                Ok(False),
+                Ok(SwimmingLevel.ADVANCED),
+                Ok(9.5),
+                Ok(3.1),
+                Ok("USA"),
+                Person,
+            )
+            == Err(("invalid first name", "invalid last name", "invalid eye color"))
+    )
+
+    assert (
+            validate_and_map(
+                Err("invalid first name"),
+                Ok("smith"),
+                Ok(25),
+                Err("invalid eye color"),
+                Ok(False),
+                Ok(SwimmingLevel.ADVANCED),
+                Ok(9.5),
+                Ok(3.1),
+                Ok("USA"),
+                Person,
+            )
+            == Err(("invalid first name", "invalid eye color"))
+    )
+
+    assert validate_and_map(
+        Ok("John"),
+        Ok("Doe"),
+        Ok(25),
+        Ok("brown"),
+        Ok(False),
+        Ok(SwimmingLevel.ADVANCED),
+        Ok(9.5),
+        Ok(3.1),
+        Ok("USA"),
+        Person,
+    ) == Ok(
+        Person(
+            first_name="John",
+            last_name="Doe",
+            age=25,
+            eye_color="brown",
+            can_fly=False,
+            swimming_level=SwimmingLevel.ADVANCED,
+            number_of_fingers=9.5,
+            number_of_toes=3.1,
+            country="USA",
+        )
+    )
+
+
+def test_map_10() -> None:
+    class SwimmingLevel(Enum):
+        BEGINNER = 1
+        INTERMEDIATE = 2
+        ADVANCED = 3
+
+    @dataclass
+    class Person:
+        first_name: str
+        last_name: str
+        age: int
+        eye_color: str
+        can_fly: bool
+        swimming_level: SwimmingLevel
+        number_of_fingers: float
+        number_of_toes: float
+        country: str
+        region: str
+
+    assert validate_and_map(
+        Err("invalid first name"),
+        Err("invalid last name"),
+        Err("invalid age"),
+        Err("invalid eye color"),
+        Err("invalid -- people can't fly"),
+        Err("invalid -- beginners not allowed"),
+        Err("invalid number of fingers"),
+        Err("invalid number of toes"),
+        Err("invalid country"),
+        Err("invalid region"),
+        Person,
+    ) == Err(
+        (
+            "invalid first name",
+            "invalid last name",
+            "invalid age",
+            "invalid eye color",
+            "invalid -- people can't fly",
+            "invalid -- beginners not allowed",
+            "invalid number of fingers",
+            "invalid number of toes",
+            "invalid country",
+            "invalid region",
+        )
+    )
+
+    assert (
+            validate_and_map(
+                Err("invalid first name"),
+                Err("invalid last name"),
+                Ok(25),
+                Err("invalid eye color"),
+                Ok(False),
+                Ok(SwimmingLevel.ADVANCED),
+                Ok(9.5),
+                Ok(3.1),
+                Ok("USA"),
+                Ok("California"),
+                Person,
+            )
+            == Err(("invalid first name", "invalid last name", "invalid eye color"))
+    )
+
+    assert (
+            validate_and_map(
+                Err("invalid first name"),
+                Ok("smith"),
+                Ok(25),
+                Err("invalid eye color"),
+                Ok(False),
+                Ok(SwimmingLevel.ADVANCED),
+                Ok(9.5),
+                Ok(3.1),
+                Ok("USA"),
+                Ok("California"),
+                Person,
+            )
+            == Err(("invalid first name", "invalid eye color"))
+    )
+
+    assert validate_and_map(
+        Ok("John"),
+        Ok("Doe"),
+        Ok(25),
+        Ok("brown"),
+        Ok(False),
+        Ok(SwimmingLevel.ADVANCED),
+        Ok(9.5),
+        Ok(3.1),
+        Ok("USA"),
+        Ok("California"),
+        Person,
+    ) == Ok(
+        Person(
+            first_name="John",
+            last_name="Doe",
+            age=25,
+            eye_color="brown",
+            can_fly=False,
+            swimming_level=SwimmingLevel.ADVANCED,
+            number_of_fingers=9.5,
+            number_of_toes=3.1,
+            country="USA",
+            region="California",
+        )
+    )
